NSObject+RACSelectorSignal: fix race condition.
The swizzling methods in NSObject+RACSelectorSignal were synchronizing
on the object's instance, but the changes they are doing are global to
the class of the object. Therefore, calling -rac_signalForSelector:
before a subclass is created and the methods are swizzled,
simultaneously, on two instances of the same object, may raise an
assert.

The solution is synchronize all class-wide changes on the class object.