RACSignal: override combinePreviousWithStart:reduce: implementation.
Current implementation of combinePreviousWithStart:reduce: in RACSignal
is originated in RACStream ancestor and relies on the method
scanWithStart:reduce:.
Using the method forces implementation to pack the current sent value
with the reduce(previous, current) value in the scan block: the current
value is used for the next run of the scan, becoming the previous value.
The reduce(previous, current) is used for forwarding the reduced value
in the pipeline. The problem with that method is that
reduce(previous, current) gets retained for no practical reason until
the next value gets sent.

The overriding implementation of the method uses a straightforward
approach without relying on other RACStream existing operators. It is
also much more intuitive for reading than the original implementation.

See [RACStream combinePreviousWithStart:reduce:] for more information.