ReactiveObjCTests: fix TSan issues.
Most of the fixes involve changing `BOOL` and `NSUInteger` to their
atomic counterparts. Since the original types are atomic on x64 and arm
and the tests only check that eventually they are equal to a specific
value (which implies relaxed ordering), no real difference in the result
of the test is expected.

There were two tests where I needed to change the logic of the test,
since the matcher was expecting a pointer which contains an ObjC value
to change, and since it cannot be expressed by an atomic variable I
changed the expected variable to an `atomic_bool`.