RACSignal: annotate -subscribe: methods to return nullable disposable.
The -subscribe: method family is annotated as returning a `nonnull`
`RACDisposable`. However, some subclasses of `RACSignal` may return a
`nil` disposable in cases where the subscription cannot be disposed of
(mostly in special signals where the subscription begins and terminates
without a way for the user to dispose the subscription in the middle).

To avoid breaking API and to affect performance, annotating the API
correctly seems like a better option than making the code return a
`RACDisposable` in all code paths.