RACNeverSignal: initial commit.
The fix for `finally:` in 45f673b surfaced an issue with `finally:`
operators that are chained to `+[RACSignal never]` signals.

Intuitively, since this signal never completes or errs, the `finally:`
block should be called only on explicit disposal of the subscription.
Instead, the block is called immediately upon subscription. This happens
since `[RACSignal never]` is implemented as an empty dynamic signal
which doesn't retain its subscriber, and thus the subscriber is
deallocated immediately, which disposes the subscription disposable and
calls the block.

The proposed solution implements a new `RACNeverSignal` which uses
`RACPassthroughSubscriber` to create a retain cycle that holds the
subscriber until the subscription is disposed.

This adds additional complexity for `+[RACSignal never]` signals, but
since these signals are used mostly in tests the memory and performance
degradations should be minimal. The potential downside in this change is
having subscriptions that were used to automatically end to leak, but
subscriptions to infinite signals that are not disposed of are already
considered a bug, so this change will only surface the issue.